Unified API almost complete (for the basic aims)
Unified API almost complete (for the basic aims): device handling,
memory allocation and copy are now unified, the seamless integration of
OpenACC and OpenMP backends is near to be completed.

A taste examples has been added and reported into the README (also
improved with some documentation).

All the tests can now be compiled with both backends, but only OpenACC
are also been executed and checked, the OpenMP compiled tests are not
yet executed on Intel GPUs.